Adds "empty" game piece as board default
Instead of an empty square returning a " " by default, it returns
the blank space, which is the parent class of all game pieces. Its
to_s returns a " ". All game piece classes now have an empty?
method that returns false if it is a real game piece, and true if
it is the blank space.

Also, Player classes have a to_s method that returns their team
color, capitalized, which simplifies string interpolation.